+++ test/src/metadata_tests.cpp
@@ -13,31 +13,11 @@ TEST(MetadataTests, TestDatasetConditionDetailToString) {
       R"(DatasetConditionDetail { date = "2022-11-10", condition = available })");
 }
 
-TEST(MetadataTests, TestDatasetConditionInfoToString) {
-  const DatasetConditionInfo target{
-      DatasetCondition::Bad,
-      {{"2022-11-07", DatasetCondition::Available},
-       {"2022-11-08", DatasetCondition::Bad},
-       {"2022-11-09", DatasetCondition::Bad},
-       {"2022-11-10", DatasetCondition::Available}},
-      "2022-11-07",
-      "2022-11-10",
-      "2017-05-21",
-      "2023-03-01"};
-  const auto res = ToString(target);
-  ASSERT_EQ(res, R"(DatasetConditionInfo {
-    condition = bad,
-    details = {
-        DatasetConditionDetail { date = "2022-11-07", condition = available },
-        DatasetConditionDetail { date = "2022-11-08", condition = bad },
-        DatasetConditionDetail { date = "2022-11-09", condition = bad },
-        DatasetConditionDetail { date = "2022-11-10", condition = available }
-    },
-    adjusted_start_date = "2022-11-07",
-    adjusted_end_date = "2022-11-10",
-    available_start_date = "2017-05-21",
-    available_end_date = "2023-03-01"
-})");
+TEST(MetadataTests, TestDatasetRangeToString) {
+  const DatasetRange target{"2022-05-17", "2023-01-07"};
+  ASSERT_EQ(
+      ToString(target),
+      R"(DatasetRange { start_date = "2022-05-17", end_date = "2023-01-07" })");
 }
 }  // namespace test
 }  // namespace databento